Organizing python tests by controllers
