+++ test-http/src/test/cve_id_tests/cve_id.py
@@ -12,19 +12,20 @@ from src.utils import (assert_contains, ok_response_contains,
 CVE_ID_URL = '/api/cve-id'
 cve_id = 'CVE-1999-0001'
 
-def test_get_cve_id(reg_user_headers):
+#### GET /cve-id/:id ####
+def test_get_cve_id():
     """ the first ID from 1999 should always exist """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=reg_user_headers
+        headers=utils.BASE_HEADERS
     )
     ok_response_contains(res, cve_id)
 
 
-def test_get_cve_id_bad_org_header(reg_user_headers):
+def test_get_cve_id_bad_org_header():
     """ unauthorized users can't get known IDs """
     uid = str(uuid.uuid4())
-    tmp = copy.deepcopy(reg_user_headers)
+    tmp = copy.deepcopy(utils.BASE_HEADERS)
     tmp['CVE-API-ORG'] = uid
     tmp['CVE-API-USER'] = uid
     res = requests.get(
@@ -36,46 +37,94 @@ def test_get_cve_id_bad_org_header(reg_user_headers):
     response_contains_json(res, 'error', 'UNAUTHORIZED')
 
 
-def test_get_cve_id_id(reg_user_headers):
+def test_get_cve_id_id():
     """ the id parameter must be a string """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=reg_user_headers
+        headers=utils.BASE_HEADERS
     )
     assert isinstance (cve_id, str)
-
     
-def test_put_cve_id_id_state(reg_user_headers):
-    """ regular user cannot update id's state """
+
+#### PUT /cve-id/:id ####
+def test_put_cve_id_id():
+    """ an id can be updated to reject or public """
     res = requests.put(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
+        params={'state':'REJECT'}
+    )
+    ok_response_contains(res, f'{cve_id} was successfully updated.')
+
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS,
         params={'state':'PUBLIC'}
     )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
+    ok_response_contains(res, f'{cve_id} was successfully updated.')
 
 
-def test_post_cve_id_update_parameters(reg_user_headers):
-    """ org users can update own information """
-    res = requests.post(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
-        params={
-            'amount': '10',
-            'batch_type': 'sequential',
-            'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': reg_user_headers['CVE-API-ORG']
-        }
+def test_put_cve_id_id_state_reserved():
+    """ an id's state cannot be set back to reserved """
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS,
+        params={'state':'REJECT'}
     )
-    assert res.status_code == 200
+    ok_response_contains(res, f'{cve_id} was successfully updated.')
+
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS,
+        params={'state':'RESERVED'}
+    )
+    assert res.status_code == 400
+    response_contains_json(
+        res, 'message',
+        'Cannot change the state to RESERVED.')
 
 
-def test_post_cve_id_no_params(reg_user_headers):
+def test_put_cve_id_id_state_blarg():
+    """ an id's state can only be set to public or reject """
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS,
+        params={'state':'BLARG'}
+    )
+    assert res.status_code == 400
+    assert 'BLARG' not in res.content.decode()
+    response_contains(res, 'state')
+    response_contains_json(res, 'message', 'Parameters were invalid')  
+
+
+def test_put_cve_id_id_no_params():
+    """ cve id endpoint ... accepts an empty query set? """
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS
+    )
+    # NOTE: should this behave as if an update was successful, when no params
+    # were included and nothing was actually updated?
+    ok_response_contains(res, f'{cve_id} was successfully updated.')
+
+
+def test_put_cve_id_id_empty_params():
+    """ cve services id update endpoint fails for empty query parameters """
+    res = requests.put(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
+        headers=utils.BASE_HEADERS,
+        params={'state': '', 'cna': ''}
+    )
+    assert res.status_code == 400
+    response_contains(res, 'state')
+    response_contains_json(res, 'message', 'Parameters were invalid') 
+
+#### POST /cve-id ####
+def test_post_cve_id_no_params():
     """ batch type is the only optional parameter for reserving ids """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers
+        headers=utils.BASE_HEADERS
     )
     assert res.status_code == 400
     response_contains(res, 'amount')
@@ -83,11 +132,11 @@ def test_post_cve_id_no_params(reg_user_headers):
     response_contains(res, 'short_name')
 
 
-def test_post_cve_id_empty_params(reg_user_headers):
+def test_post_cve_id_empty_params():
     """ cve services doesn't accept id reservation with blank parameters """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '',
             'batch_type': '',
@@ -101,29 +150,59 @@ def test_post_cve_id_empty_params(reg_user_headers):
     response_contains(res, 'cve_year')
 
 
-def test_post_cve_id_wrong_header(reg_user_headers):
-    """ reg_user_headers cannot post for 'mitre' org """
+def test_post_cve_id_empty_short_name():
+    """ cve services rejects empty short name separately """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '10',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'mitre'
+            'short_name': ''
         }
     )
-    # NOTE: this error also occurs when short_name is empty and invalid, 
-    # expected error is 'NO_ORG_SHORTNAME' and 'ORG_DNE' respectively
+    assert res.status_code == 400
+    response_contains_json(res, 'error', 'NO_ORG_SHORTNAME')
+
+
+def test_post_cve_id_organization_access():
+    """ services rejects ID reservation for nonexistent Org as identified by short name """
+    res = requests.post(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
+        headers=utils.BASE_HEADERS,
+        params={
+            'amount': '10',
+            'batch_type': 'sequential',
+            'cve_year': f'{utils.CURRENT_YEAR}',
+            'short_name': 'test'
+        }
+    )
     assert res.status_code == 403
-    response_contains_json(res, 'error', 'ORG_CANNOT_RESERVE_FOR_OTHER')
+    response_contains_json(res, 'error', 'ORG_DNE') 
 
 
-def test_post_cve_id_empty_year(reg_user_headers):
+def test_post_cve_id_shortname_format():
+    """ cve services rejects shortname that's not a string """
+    res = requests.post(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
+        headers=utils.BASE_HEADERS,
+        params={
+            'amount': '10',
+            'batch_type': 'sequential',
+            'cve_year': f'{utils.CURRENT_YEAR}',
+            'short_name': '1111'
+        }
+    )
+    assert res.status_code == 403
+    response_contains_json(res, 'error', 'ORG_DNE') 
+
+
+def test_post_cve_id_empty_year():
     """ cve services rejects empty year """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -135,11 +214,11 @@ def test_post_cve_id_empty_year(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_bad_year(reg_user_headers):
+def test_post_cve_id_bad_year():
     """ cve services rejects year that isn't a 4 digit number"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -151,11 +230,11 @@ def test_post_cve_id_bad_year(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_post_cve_id_empty_amount(reg_user_headers):
+def test_post_cve_id_empty_amount():
     """ cve services rejects empty amount """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '',
             'batch_type': 'sequential',
@@ -167,72 +246,87 @@ def test_post_cve_id_empty_amount(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_invalid_amount(reg_user_headers):
+def test_post_cve_id_invalid_amount():
     """ cve services rejects amount less than or equal to 0 """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '-1',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': reg_user_headers['CVE-API-ORG']
+            'short_name': 'mitre'
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_AMOUNT')
 
 
-def test_post_cve_id_no_batch_type(reg_user_headers):
+def test_post_cve_id_no_batch_type():
     """ cve services rejects not having a batch type"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '10',
             'batch_type': '',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': reg_user_headers['CVE-API-ORG']
+            'short_name': 'mitre'
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'NO_BATCH_TYPE')
 
 
-def test_post_cve_id_invalid_batch_type(reg_user_headers):
+def test_post_cve_id_invalid_batch_type():
     """ cve services rejects batch types that aren't 'sequential' or 'nonsequential' """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '10',
             'batch_type': '---',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name':  reg_user_headers['CVE-API-ORG']
+            'short_name': 'mitre'
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_BATCH_TYPE')
 
 
-def test_post_cve_id_bad_amount(reg_user_headers):
+def test_post_cve_id_bad_amount():
     """ api rejects non-numeric amount when requesting IDs """
-    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, reg_user_headers['CVE-API-USER'])
+    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, 'mitre')
     assert res.status_code == 400
     assert res.reason == 'Bad Request'
     response_contains_json(res, 'error', 'BAD_INPUT')
     assert_contains(res, 'amount')
 
 
-def test_post_cve_id_reserve_priority(reg_user_headers):
-    """ priority ids can be reserved on behalf of the general user org """
+#### PUT /cve-id/:id ####
+def test_put_cve_id_duplicate_state():
+    """ api rejects ID updates with duplicate state parameter """
+    res = requests.put(
+        (f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}'
+          '?state=RESERVED&state=PUBLIC'),
+        headers=utils.BASE_HEADERS
+    )
+    assert res.status_code == 400
+    assert res.reason == 'Bad Request'
+    response_contains_json(res, 'error', 'BAD_INPUT')
+    assert_contains(res, 'state')
+
+
+#### POST /cve-id ####
+def test_post_cve_id_reserve_priority():
+    """ priority ids can be reserved on behalf of the mitre org """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'amount': '1',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': reg_user_headers['CVE-API-ORG']
+            'short_name': 'mitre'
         }
     )
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
@@ -250,9 +344,9 @@ def test_post_cve_id_reserve_priority(reg_user_headers):
     "batch_type, amount",
     [('sequential', 10), ('sequential', 1000),
      ('nonsequential', 1), ('nonsequential', 10)])
-def test_post_cve_id_reservation(batch_type, amount, reg_user_headers):
+def test_post_cve_id_reservation(batch_type, amount):
     """ sequential ids can be reserved on behalf of the mitre org """
-    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'], batch_type)
+    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, 'mitre', batch_type)
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
     assert json.loads(res.content.decode())['cve_ids']
     assert len(json.loads(res.content.decode())['cve_ids']) == amount
@@ -261,21 +355,22 @@ def test_post_cve_id_reservation(batch_type, amount, reg_user_headers):
     assert 'user' in res.content.decode()
     
 
-def test_post_cve_id_reserve_sequential_over_quota(reg_user_headers):
+def test_post_cve_id_reserve_sequential_over_quota():
     """ the services api enforces a max quota of 100,000 """
-    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
+    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, 'mitre')
     assert res.status_code == 403
     response_contains_json(res, 'error', 'EXCEEDED_ID_QUOTA')
 
 
-def test_post_cve_id_reserve_nonsequential_over_limit(reg_user_headers):
+def test_post_cve_id_reserve_nonsequential_over_limit():
     """ the services api enforces a max non-sequential limit of 10 """
-    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'], 'nonsequential')
+    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, 'mitre', 'nonsequential')
     assert res.status_code == 403
     response_contains_json(res, 'error', 'OVER_NONSEQUENTIAL_MAX_AMOUNT')
 
 
-def test_get_cve_id_by_time_reserved(reg_user_headers):
+#### GET /cve-id ####
+def test_get_cve_id_by_time_reserved():
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -283,7 +378,7 @@ def test_get_cve_id_by_time_reserved(reg_user_headers):
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, 'mitre')
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -299,7 +394,7 @@ def test_get_cve_id_by_time_reserved(reg_user_headers):
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_by_time_modified(reg_user_headers):
+def test_get_cve_id_by_time_modified():
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -307,7 +402,7 @@ def test_get_cve_id_by_time_modified(reg_user_headers):
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, 'mitre')
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -323,25 +418,11 @@ def test_get_cve_id_by_time_modified(reg_user_headers):
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_with_params(reg_user_headers):
-    """ org user can retrieve ids"""
-    res = requests.get(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
-        params={
-            'page': 1,
-            'state': 'PUBLIC',
-            'cve_id_year': 2011
-        }
-    )
-    assert res.status_code == 200
-
-
-def test_get_cve_id_empty_parameters(reg_user_headers):
+def test_get_cve_id_empty_parameters():
     """ cannot get id with empty parameters """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'page': ' ',
             'state': ' ',
@@ -363,11 +444,11 @@ def test_get_cve_id_empty_parameters(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_format_number(reg_user_headers):
+def test_get_cve_id_page_format_number():
     """ page must be an integer' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'page': 'test',
         }
@@ -376,11 +457,11 @@ def test_get_cve_id_page_format_number(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_limit(reg_user_headers):
+def test_get_cve_id_page_limit():
     """ page must be greater than or equal to 1' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'page': '-1',
         }
@@ -389,11 +470,11 @@ def test_get_cve_id_page_limit(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_state_in_choices(reg_user_headers):
+def test_get_cve_id_state_in_choices():
     """ state parameter can only be 'REJECT', 'PUBLIC' or 'RESERVED' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'state': 'TEST',
         }
@@ -402,11 +483,11 @@ def test_get_cve_id_state_in_choices(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_letters(reg_user_headers):
+def test_get_cve_id_year_format_with_letters():
     """ cve_id_year format cannot have letters  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'cve_id_year': 'test',
         }
@@ -415,11 +496,11 @@ def test_get_cve_id_year_format_with_letters(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_digits(reg_user_headers):
+def test_get_cve_id_year_format_with_digits():
     """ cve_id_year format must be 4 digits only  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'cve_id_year': '20111',
         }
@@ -428,11 +509,11 @@ def test_get_cve_id_year_format_with_digits(reg_user_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
     
 
-def test_get_cve_id_available_state(reg_user_headers):
+def test_get_cve_id_available_state():
     """ CVE ID filter endpoint does not return any IDs with state 'AVAILABLE' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=reg_user_headers,
+        headers=utils.BASE_HEADERS,
         params={
             'page': 1,
             'state': 'PUBLIC',
@@ -441,6 +522,7 @@ def test_get_cve_id_available_state(reg_user_headers):
     )
     assert res.status_code == 200
     assert 'AVAILABLE' not in res.content.decode()
+    response_contains(res, 'state')
  
 
 # CVE-ID ENDPOINT UTILITIES
