+++ test-http/src/test/cve_id_tests/cve_id_as_general_user.py
@@ -12,19 +12,20 @@ from src.utils import (assert_contains, ok_response_contains,
 CVE_ID_URL = '/api/cve-id'
 cve_id = 'CVE-1999-0001'
 
-def test_get_cve_id(org_admin_headers):
+#### GET /cve-id/:id ####
+def test_get_cve_id(reg_user_headers):
     """ the first ID from 1999 should always exist """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=org_admin_headers
+        headers=reg_user_headers
     )
     ok_response_contains(res, cve_id)
 
 
-def test_get_cve_id_bad_org_header(org_admin_headers):
+def test_get_cve_id_bad_org_header(reg_user_headers):
     """ unauthorized users can't get known IDs """
     uid = str(uuid.uuid4())
-    tmp = copy.deepcopy(org_admin_headers)
+    tmp = copy.deepcopy(reg_user_headers)
     tmp['CVE-API-ORG'] = uid
     tmp['CVE-API-USER'] = uid
     res = requests.get(
@@ -36,46 +37,48 @@ def test_get_cve_id_bad_org_header(org_admin_headers):
     response_contains_json(res, 'error', 'UNAUTHORIZED')
 
 
-def test_get_cve_id_id(org_admin_headers):
+def test_get_cve_id_id(reg_user_headers):
     """ the id parameter must be a string """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=org_admin_headers
+        headers=reg_user_headers
     )
     assert isinstance (cve_id, str)
 
-    
-def test_put_cve_id_id_state(org_admin_headers):
-    """ org admin cannot update id's state """
+
+#### PUT /cve-id/:id ####
+def test_put_cve_id_id_state(reg_user_headers):
+    """ regular user cannot update id's state """
     res = requests.put(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={'state':'PUBLIC'}
     )
     assert res.status_code == 403
     response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
-def test_post_cve_id_update_parameters(org_admin_headers):
-    """ org admins can update own information """
+#### POST /cve-id ####
+def test_post_cve_id_update_parameters(reg_user_headers):
+    """ org users can update own information """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': org_admin_headers['CVE-API-ORG']
+            'short_name': reg_user_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 200
 
 
-def test_post_cve_id_no_params(org_admin_headers):
+def test_post_cve_id_no_params(reg_user_headers):
     """ batch type is the only optional parameter for reserving ids """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers
+        headers=reg_user_headers
     )
     assert res.status_code == 400
     response_contains(res, 'amount')
@@ -83,11 +86,11 @@ def test_post_cve_id_no_params(org_admin_headers):
     response_contains(res, 'short_name')
 
 
-def test_post_cve_id_empty_params(org_admin_headers):
+def test_post_cve_id_empty_params(reg_user_headers):
     """ cve services doesn't accept id reservation with blank parameters """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '',
             'batch_type': '',
@@ -101,11 +104,11 @@ def test_post_cve_id_empty_params(org_admin_headers):
     response_contains(res, 'cve_year')
 
 
-def test_post_cve_id_wrong_header(org_admin_headers):
-    """ org_admin_headers cannot post for 'mitre' org """
+def test_post_cve_id_wrong_header(reg_user_headers):
+    """ reg_user_headers cannot post for 'mitre' org """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -113,17 +116,17 @@ def test_post_cve_id_wrong_header(org_admin_headers):
             'short_name': 'mitre'
         }
     )
-    # NOTE: this error also occurs when short_name is empty, expected error is 'NO_ORG_SHORTNAME'
-    # and when short_name is invalid, expected error is 'ORG_DNE'
+    # NOTE: this error also occurs when short_name is empty and invalid, 
+    # expected error is 'NO_ORG_SHORTNAME' and 'ORG_DNE' respectively
     assert res.status_code == 403
     response_contains_json(res, 'error', 'ORG_CANNOT_RESERVE_FOR_OTHER')
 
 
-def test_post_cve_id_empty_year(org_admin_headers):
+def test_post_cve_id_empty_year(reg_user_headers):
     """ cve services rejects empty year """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -135,11 +138,11 @@ def test_post_cve_id_empty_year(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_bad_year(org_admin_headers):
+def test_post_cve_id_bad_year(reg_user_headers):
     """ cve services rejects year that isn't a 4 digit number"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -151,11 +154,11 @@ def test_post_cve_id_bad_year(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_post_cve_id_empty_amount(org_admin_headers):
+def test_post_cve_id_empty_amount(reg_user_headers):
     """ cve services rejects empty amount """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '',
             'batch_type': 'sequential',
@@ -167,72 +170,72 @@ def test_post_cve_id_empty_amount(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_invalid_amount(org_admin_headers):
+def test_post_cve_id_invalid_amount(reg_user_headers):
     """ cve services rejects amount less than or equal to 0 """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '-1',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': org_admin_headers['CVE-API-ORG']
+            'short_name': reg_user_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_AMOUNT')
 
-  
-def test_post_cve_id_no_batch_type(org_admin_headers):
+
+def test_post_cve_id_no_batch_type(reg_user_headers):
     """ cve services rejects not having a batch type"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': '',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': org_admin_headers['CVE-API-ORG']
+            'short_name': reg_user_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'NO_BATCH_TYPE')
 
 
-def test_post_cve_id_invalid_batch_type(org_admin_headers):
+def test_post_cve_id_invalid_batch_type(reg_user_headers):
     """ cve services rejects batch types that aren't 'sequential' or 'nonsequential' """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '10',
             'batch_type': '---',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': org_admin_headers['CVE-API-ORG']
+            'short_name':  reg_user_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_BATCH_TYPE')
 
 
-def test_post_cve_id_bad_amount(org_admin_headers):
+def test_post_cve_id_bad_amount(reg_user_headers):
     """ api rejects non-numeric amount when requesting IDs """
-    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
+    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, reg_user_headers['CVE-API-USER'])
     assert res.status_code == 400
     assert res.reason == 'Bad Request'
     response_contains_json(res, 'error', 'BAD_INPUT')
     assert_contains(res, 'amount')
 
 
-def test_post_cve_id_reserve_priority(org_admin_headers):
-    """ priority ids can be reserved on behalf of the admin's org """
+def test_post_cve_id_reserve_priority(reg_user_headers):
+    """ priority ids can be reserved on behalf of the general user org """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'amount': '1',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': org_admin_headers['CVE-API-ORG']
+            'short_name': reg_user_headers['CVE-API-ORG']
         }
     )
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
@@ -250,9 +253,9 @@ def test_post_cve_id_reserve_priority(org_admin_headers):
     "batch_type, amount",
     [('sequential', 10), ('sequential', 1000),
      ('nonsequential', 1), ('nonsequential', 10)])
-def test_post_cve_id_reservation(batch_type, amount, org_admin_headers):
+def test_post_cve_id_reservation(batch_type, amount, reg_user_headers):
     """ sequential ids can be reserved on behalf of the mitre org """
-    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'], batch_type)
+    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'], batch_type)
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
     assert json.loads(res.content.decode())['cve_ids']
     assert len(json.loads(res.content.decode())['cve_ids']) == amount
@@ -261,21 +264,22 @@ def test_post_cve_id_reservation(batch_type, amount, org_admin_headers):
     assert 'user' in res.content.decode()
     
 
-def test_post_cve_id_reserve_sequential_over_quota(org_admin_headers):
+def test_post_cve_id_reserve_sequential_over_quota(reg_user_headers):
     """ the services api enforces a max quota of 100,000 """
-    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
+    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
     assert res.status_code == 403
     response_contains_json(res, 'error', 'EXCEEDED_ID_QUOTA')
 
 
-def test_post_cve_id_reserve_nonsequential_over_limit(org_admin_headers):
+def test_post_cve_id_reserve_nonsequential_over_limit(reg_user_headers):
     """ the services api enforces a max non-sequential limit of 10 """
-    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'], 'nonsequential')
+    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'], 'nonsequential')
     assert res.status_code == 403
     response_contains_json(res, 'error', 'OVER_NONSEQUENTIAL_MAX_AMOUNT')
 
 
-def test_get_cve_id_by_time_reserved(org_admin_headers):
+#### GET /cve-id ####
+def test_get_cve_id_by_time_reserved(reg_user_headers):
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -283,7 +287,7 @@ def test_get_cve_id_by_time_reserved(org_admin_headers):
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -299,7 +303,7 @@ def test_get_cve_id_by_time_reserved(org_admin_headers):
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_by_time_modified(org_admin_headers):
+def test_get_cve_id_by_time_modified(reg_user_headers):
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -307,7 +311,7 @@ def test_get_cve_id_by_time_modified(org_admin_headers):
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, reg_user_headers['CVE-API-ORG'])
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -323,11 +327,11 @@ def test_get_cve_id_by_time_modified(org_admin_headers):
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_with_params(org_admin_headers):
-    """ org admin can retrieve ids"""
+def test_get_cve_id_with_params(reg_user_headers):
+    """ org user can retrieve ids"""
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'page': 1,
             'state': 'PUBLIC',
@@ -337,11 +341,11 @@ def test_get_cve_id_with_params(org_admin_headers):
     assert res.status_code == 200
 
 
-def test_get_cve_id_empty_parameters(org_admin_headers):
+def test_get_cve_id_empty_parameters(reg_user_headers):
     """ cannot get id with empty parameters """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'page': ' ',
             'state': ' ',
@@ -363,11 +367,11 @@ def test_get_cve_id_empty_parameters(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_format_number(org_admin_headers):
+def test_get_cve_id_page_format_number(reg_user_headers):
     """ page must be an integer' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'page': 'test',
         }
@@ -376,11 +380,11 @@ def test_get_cve_id_page_format_number(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_limit(org_admin_headers):
+def test_get_cve_id_page_limit(reg_user_headers):
     """ page must be greater than or equal to 1' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'page': '-1',
         }
@@ -389,11 +393,11 @@ def test_get_cve_id_page_limit(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_state_in_choices(org_admin_headers):
+def test_get_cve_id_state_in_choices(reg_user_headers):
     """ state parameter can only be 'REJECT', 'PUBLIC' or 'RESERVED' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'state': 'TEST',
         }
@@ -402,11 +406,11 @@ def test_get_cve_id_state_in_choices(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_letters(org_admin_headers):
+def test_get_cve_id_year_format_with_letters(reg_user_headers):
     """ cve_id_year format cannot have letters  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'cve_id_year': 'test',
         }
@@ -415,11 +419,11 @@ def test_get_cve_id_year_format_with_letters(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_digits(org_admin_headers):
+def test_get_cve_id_year_format_with_digits(reg_user_headers):
     """ cve_id_year format must be 4 digits only  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'cve_id_year': '20111',
         }
@@ -428,11 +432,11 @@ def test_get_cve_id_year_format_with_digits(org_admin_headers):
     response_contains_json(res, 'error', 'BAD_INPUT')
     
 
-def test_get_cve_id_available_state(org_admin_headers):
+def test_get_cve_id_available_state(reg_user_headers):
     """ CVE ID filter endpoint does not return any IDs with state 'AVAILABLE' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=org_admin_headers,
+        headers=reg_user_headers,
         params={
             'page': 1,
             'state': 'PUBLIC',
@@ -441,7 +445,7 @@ def test_get_cve_id_available_state(org_admin_headers):
     )
     assert res.status_code == 200
     assert 'AVAILABLE' not in res.content.decode()
-
+ 
 
 # CVE-ID ENDPOINT UTILITIES
 # ==============================================================================
