+++ test-http/src/test/cve_id_tests/cve_id_as_org_admin.py
@@ -13,19 +13,20 @@ CVE_ID_URL = '/api/cve-id'
 cve_id = 'CVE-1999-0001'
 
 
-def test_get_cve_id():
+#### GET /cve-id:id ####
+def test_get_cve_id(org_admin_headers):
     """ the first ID from 1999 should always exist """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS
+        headers=org_admin_headers
     )
     ok_response_contains(res, cve_id)
 
 
-def test_get_cve_id_bad_org_header():
+def test_get_cve_id_bad_org_header(org_admin_headers):
     """ unauthorized users can't get known IDs """
     uid = str(uuid.uuid4())
-    tmp = copy.deepcopy(utils.BASE_HEADERS)
+    tmp = copy.deepcopy(org_admin_headers)
     tmp['CVE-API-ORG'] = uid
     tmp['CVE-API-USER'] = uid
     res = requests.get(
@@ -37,93 +38,48 @@ def test_get_cve_id_bad_org_header():
     response_contains_json(res, 'error', 'UNAUTHORIZED')
 
 
-def test_get_cve_id_id():
+def test_get_cve_id_id(org_admin_headers):
     """ the id parameter must be a string """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS
+        headers=org_admin_headers
     )
     assert isinstance (cve_id, str)
-    
 
-def test_put_cve_id_id():
-    """ an id can be updated to reject or public """
+
+#### PUT /cve-id:id ####
+def test_put_cve_id_id_state(org_admin_headers):
+    """ org admin cannot update id's state """
     res = requests.put(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
-        params={'state':'REJECT'}
-    )
-    ok_response_contains(res, f'{cve_id} was successfully updated.')
-
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={'state':'PUBLIC'}
     )
-    ok_response_contains(res, f'{cve_id} was successfully updated.')
+    assert res.status_code == 403
+    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
-def test_put_cve_id_id_state_reserved():
-    """ an id's state cannot be set back to reserved """
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
-        params={'state':'REJECT'}
+#### POST /cve-id ####
+def test_post_cve_id_update_parameters(org_admin_headers):
+    """ org admins can update own information """
+    res = requests.post(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
+        headers=org_admin_headers,
+        params={
+            'amount': '10',
+            'batch_type': 'sequential',
+            'cve_year': f'{utils.CURRENT_YEAR}',
+            'short_name': org_admin_headers['CVE-API-ORG']
+        }
     )
-    ok_response_contains(res, f'{cve_id} was successfully updated.')
-
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
-        params={'state':'RESERVED'}
-    )
-    assert res.status_code == 400
-    response_contains_json(
-        res, 'message',
-        'Cannot change the state to RESERVED.')
+    assert res.status_code == 200
 
 
-def test_put_cve_id_id_state_blarg():
-    """ an id's state can only be set to public or reject """
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
-        params={'state':'BLARG'}
-    )
-    assert res.status_code == 400
-    assert 'BLARG' not in res.content.decode()
-    response_contains(res, 'state')
-    response_contains_json(res, 'message', 'Parameters were invalid')  
-
-
-def test_put_cve_id_id_no_params():
-    """ cve id endpoint ... accepts an empty query set? """
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS
-    )
-    # NOTE: should this behave as if an update was successful, when no params
-    # were included and nothing was actually updated?
-    ok_response_contains(res, f'{cve_id} was successfully updated.')
-
-
-def test_put_cve_id_id_empty_params():
-    """ cve services id update endpoint fails for empty query parameters """
-    res = requests.put(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}',
-        headers=utils.BASE_HEADERS,
-        params={'state': '', 'cna': ''}
-    )
-    assert res.status_code == 400
-    response_contains(res, 'state')
-    response_contains_json(res, 'message', 'Parameters were invalid') 
-
-
-def test_post_cve_id_no_params():
+def test_post_cve_id_no_params(org_admin_headers):
     """ batch type is the only optional parameter for reserving ids """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS
+        headers=org_admin_headers
     )
     assert res.status_code == 400
     response_contains(res, 'amount')
@@ -131,11 +87,11 @@ def test_post_cve_id_no_params():
     response_contains(res, 'short_name')
 
 
-def test_post_cve_id_empty_params():
+def test_post_cve_id_empty_params(org_admin_headers):
     """ cve services doesn't accept id reservation with blank parameters """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '',
             'batch_type': '',
@@ -149,59 +105,29 @@ def test_post_cve_id_empty_params():
     response_contains(res, 'cve_year')
 
 
-def test_post_cve_id_empty_short_name():
-    """ cve services rejects empty short name separately """
+def test_post_cve_id_wrong_header(org_admin_headers):
+    """ org_admin_headers cannot post for 'mitre' org """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': ''
+            'short_name': 'mitre'
         }
     )
-    assert res.status_code == 400
-    response_contains_json(res, 'error', 'NO_ORG_SHORTNAME')
-
-
-def test_post_cve_id_organization_access():
-    """ services rejects ID reservation for nonexistent Org as identified by short name """
-    res = requests.post(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
-        params={
-            'amount': '10',
-            'batch_type': 'sequential',
-            'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'test'
-        }
-    )
+    # NOTE: this error also occurs when short_name is empty, expected error is 'NO_ORG_SHORTNAME'
+    # and when short_name is invalid, expected error is 'ORG_DNE'
     assert res.status_code == 403
-    response_contains_json(res, 'error', 'ORG_DNE') 
+    response_contains_json(res, 'error', 'ORG_CANNOT_RESERVE_FOR_OTHER')
 
 
-def test_post_cve_id_shortname_format():
-    """ cve services rejects shortname that's not a string """
-    res = requests.post(
-        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
-        params={
-            'amount': '10',
-            'batch_type': 'sequential',
-            'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': '1111'
-        }
-    )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'ORG_DNE') 
-
-
-def test_post_cve_id_empty_year():
+def test_post_cve_id_empty_year(org_admin_headers):
     """ cve services rejects empty year """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -213,11 +139,11 @@ def test_post_cve_id_empty_year():
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_bad_year():
+def test_post_cve_id_bad_year(org_admin_headers):
     """ cve services rejects year that isn't a 4 digit number"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '10',
             'batch_type': 'sequential',
@@ -229,11 +155,11 @@ def test_post_cve_id_bad_year():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_post_cve_id_empty_amount():
+def test_post_cve_id_empty_amount(org_admin_headers):
     """ cve services rejects empty amount """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '',
             'batch_type': 'sequential',
@@ -245,85 +171,72 @@ def test_post_cve_id_empty_amount():
     response_contains_json(res, 'error', 'BAD_INPUT') 
 
 
-def test_post_cve_id_invalid_amount():
+def test_post_cve_id_invalid_amount(org_admin_headers):
     """ cve services rejects amount less than or equal to 0 """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '-1',
             'batch_type': 'sequential',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'mitre'
+            'short_name': org_admin_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_AMOUNT')
 
-
-def test_post_cve_id_no_batch_type():
+  
+def test_post_cve_id_no_batch_type(org_admin_headers):
     """ cve services rejects not having a batch type"""
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '10',
             'batch_type': '',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'mitre'
+            'short_name': org_admin_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'NO_BATCH_TYPE')
 
 
-def test_post_cve_id_invalid_batch_type():
+def test_post_cve_id_invalid_batch_type(org_admin_headers):
     """ cve services rejects batch types that aren't 'sequential' or 'nonsequential' """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '10',
             'batch_type': '---',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'mitre'
+            'short_name': org_admin_headers['CVE-API-ORG']
         }
     )
     assert res.status_code == 400
     response_contains_json(res, 'error', 'INVALID_BATCH_TYPE')
 
 
-def test_post_cve_id_bad_amount():
+def test_post_cve_id_bad_amount(org_admin_headers):
     """ api rejects non-numeric amount when requesting IDs """
-    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, 'mitre')
+    res = get_reserve_cve_ids('a', utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
     assert res.status_code == 400
     assert res.reason == 'Bad Request'
     response_contains_json(res, 'error', 'BAD_INPUT')
     assert_contains(res, 'amount')
 
 
-def test_put_cve_id_duplicate_state():
-    """ api rejects ID updates with duplicate state parameter """
-    res = requests.put(
-        (f'{env.AWG_BASE_URL}{CVE_ID_URL}/{cve_id}'
-          '?state=RESERVED&state=PUBLIC'),
-        headers=utils.BASE_HEADERS
-    )
-    assert res.status_code == 400
-    assert res.reason == 'Bad Request'
-    response_contains_json(res, 'error', 'BAD_INPUT')
-    assert_contains(res, 'state')
-
-
-def test_post_cve_id_reserve_priority():
-    """ priority ids can be reserved on behalf of the mitre org """
+def test_post_cve_id_reserve_priority(org_admin_headers):
+    """ priority ids can be reserved on behalf of the admin's org """
     res = requests.post(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'amount': '1',
             'cve_year': f'{utils.CURRENT_YEAR}',
-            'short_name': 'mitre'
+            'short_name': org_admin_headers['CVE-API-ORG']
         }
     )
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
@@ -341,9 +254,9 @@ def test_post_cve_id_reserve_priority():
     "batch_type, amount",
     [('sequential', 10), ('sequential', 1000),
      ('nonsequential', 1), ('nonsequential', 10)])
-def test_post_cve_id_reservation(batch_type, amount):
+def test_post_cve_id_reservation(batch_type, amount, org_admin_headers):
     """ sequential ids can be reserved on behalf of the mitre org """
-    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, 'mitre', batch_type)
+    res = get_reserve_cve_ids(amount, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'], batch_type)
     ok_response_contains(res, f'CVE-{utils.CURRENT_YEAR}-')
     assert json.loads(res.content.decode())['cve_ids']
     assert len(json.loads(res.content.decode())['cve_ids']) == amount
@@ -352,21 +265,22 @@ def test_post_cve_id_reservation(batch_type, amount):
     assert 'user' in res.content.decode()
     
 
-def test_post_cve_id_reserve_sequential_over_quota():
+def test_post_cve_id_reserve_sequential_over_quota(org_admin_headers):
     """ the services api enforces a max quota of 100,000 """
-    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, 'mitre')
+    res = get_reserve_cve_ids(100001, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
     assert res.status_code == 403
     response_contains_json(res, 'error', 'EXCEEDED_ID_QUOTA')
 
 
-def test_post_cve_id_reserve_nonsequential_over_limit():
+def test_post_cve_id_reserve_nonsequential_over_limit(org_admin_headers):
     """ the services api enforces a max non-sequential limit of 10 """
-    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, 'mitre', 'nonsequential')
+    res = get_reserve_cve_ids(11, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'], 'nonsequential')
     assert res.status_code == 403
     response_contains_json(res, 'error', 'OVER_NONSEQUENTIAL_MAX_AMOUNT')
 
 
-def test_get_cve_id_by_time_reserved():
+#### GET /cve-id ####
+def test_get_cve_id_by_time_reserved(org_admin_headers):
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -374,7 +288,7 @@ def test_get_cve_id_by_time_reserved():
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, 'mitre')
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -390,7 +304,7 @@ def test_get_cve_id_by_time_reserved():
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_by_time_modified():
+def test_get_cve_id_by_time_modified(org_admin_headers):
     """ we can get ids immediately after reserving them using the time they're
     reserved (noting that this may not work against a shared integration
     environment, we check that at least this many have been reserved) """
@@ -398,7 +312,7 @@ def test_get_cve_id_by_time_modified():
     time.sleep(1)
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
-    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, 'mitre')
+    res_ids = get_reserve_cve_ids(n_ids, utils.CURRENT_YEAR, org_admin_headers['CVE-API-ORG'])
     time.sleep(1)
     t_after = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
 
@@ -414,11 +328,25 @@ def test_get_cve_id_by_time_modified():
     assert len(json.loads(res_get_ids.content.decode())['cve_ids']) == n_ids
 
 
-def test_get_cve_id_empty_parameters():
+def test_get_cve_id_with_params(org_admin_headers):
+    """ org admin can retrieve ids"""
+    res = requests.get(
+        f'{env.AWG_BASE_URL}{CVE_ID_URL}',
+        headers=org_admin_headers,
+        params={
+            'page': 1,
+            'state': 'PUBLIC',
+            'cve_id_year': 2011
+        }
+    )
+    assert res.status_code == 200
+
+
+def test_get_cve_id_empty_parameters(org_admin_headers):
     """ cannot get id with empty parameters """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'page': ' ',
             'state': ' ',
@@ -440,11 +368,11 @@ def test_get_cve_id_empty_parameters():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_format_number():
+def test_get_cve_id_page_format_number(org_admin_headers):
     """ page must be an integer' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'page': 'test',
         }
@@ -453,11 +381,11 @@ def test_get_cve_id_page_format_number():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_page_limit():
+def test_get_cve_id_page_limit(org_admin_headers):
     """ page must be greater than or equal to 1' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'page': '-1',
         }
@@ -466,11 +394,11 @@ def test_get_cve_id_page_limit():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_state_in_choices():
+def test_get_cve_id_state_in_choices(org_admin_headers):
     """ state parameter can only be 'REJECT', 'PUBLIC' or 'RESERVED' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'state': 'TEST',
         }
@@ -479,11 +407,11 @@ def test_get_cve_id_state_in_choices():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_letters():
+def test_get_cve_id_year_format_with_letters(org_admin_headers):
     """ cve_id_year format cannot have letters  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'cve_id_year': 'test',
         }
@@ -492,11 +420,11 @@ def test_get_cve_id_year_format_with_letters():
     response_contains_json(res, 'error', 'BAD_INPUT')
 
 
-def test_get_cve_id_year_format_with_digits():
+def test_get_cve_id_year_format_with_digits(org_admin_headers):
     """ cve_id_year format must be 4 digits only  """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'cve_id_year': '20111',
         }
@@ -505,11 +433,11 @@ def test_get_cve_id_year_format_with_digits():
     response_contains_json(res, 'error', 'BAD_INPUT')
     
 
-def test_get_cve_id_available_state():
+def test_get_cve_id_available_state(org_admin_headers):
     """ CVE ID filter endpoint does not return any IDs with state 'AVAILABLE' """
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_ID_URL}',
-        headers=utils.BASE_HEADERS,
+        headers=org_admin_headers,
         params={
             'page': 1,
             'state': 'PUBLIC',
@@ -518,8 +446,7 @@ def test_get_cve_id_available_state():
     )
     assert res.status_code == 200
     assert 'AVAILABLE' not in res.content.decode()
-    response_contains(res, 'state')
- 
+
 
 # CVE-ID ENDPOINT UTILITIES
 # ==============================================================================
