+++ test-http/src/test/cve_id_tests/cve_id_performance.py
@@ -5,13 +5,13 @@ import requests
 import uuid
 from src import env, utils
 from src.utils import response_contains_json
-from src.test.cve_id import get_reserve_cve_ids
-from src.test.org import get_org_id_data, post_new_org
+from src.test.cve_id_tests.cve_id import get_reserve_cve_ids
+from src.test.org_user_tests.org import get_org_id_data, post_new_org
 
 N_JOBS = multiprocessing.cpu_count()
 ID_QUOTA_LIMIT = 100000  # FIXME: where do I belong?
 
-
+#### POST /cve-id ####
 def test_sequential_id_reservation_limit():
     """ there's no sequential reservation limit, and its been surprisingly easy
     to reserve 100,000 IDs in the past """
