+++ test-http/src/test/cve_id_tests/cve_id_range.py
@@ -6,6 +6,7 @@ from src.utils import response_contains, response_contains_json
 CVE_ID_RANGE_URL = '/api/cve-id-range'
 
 
+#### POST /cve-id-range/:year ####
 @pytest.mark.parametrize(
     "choose_year",
     [str(x) for x in range(1999, utils.CURRENT_YEAR + 1)])
