+++ test-http/src/test/cve_tests/cve.py
@@ -16,7 +16,7 @@ reserved_cve_id = 'CVE-2017-5833'
 reject_cve_id = 'CVE-2009-0009'
 
 
-# GET /cve
+#### GET /cve ####
 def test_get_all_cves():
     """ services api accepts requests for all CVEs for secretariat users """
     res = requests.get(
@@ -27,7 +27,6 @@ def test_get_all_cves():
     assert len(json.loads(res.content.decode())['cveRecords']) >= 1
 
 
-# GET /cve
 def test_get_cve_by_time_modified():
     t_before = dt.datetime.now().strftime('%Y-%m-%dT%H:%M:%S')
     time.sleep(1)
@@ -49,7 +48,7 @@ def test_get_cve_by_time_modified():
     assert len(json.loads(res.content.decode())['cveRecords']) >= 2
 
 
-# GET /cve/:id
+#### GET /cve/:id ####
 def test_get_cve():
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id}',
@@ -58,7 +57,6 @@ def test_get_cve():
     ok_response_contains(res, cve_id)
 
 
-# GET /cve/:id
 def test_get_cve_dne():
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id_dne}',
@@ -68,10 +66,10 @@ def test_get_cve_dne():
     response_contains_json(res, 'error', 'CVE_RECORD_DNE')
 
 
-# POST /cve/:id
+#### POST /cve/:id ####
 def test_create_cve():
     """ publish a cve in the PUBLIC state """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2000-0008_public.json') as json_file:
         data = json.load(json_file)
         res = requests.post(
             f'{env.AWG_BASE_URL}{CVE_URL}/{create_cve_id}',
@@ -85,10 +83,9 @@ def test_create_cve():
             response_contains_json(res, 'error', 'CVE_RECORD_EXISTS')
 
 
-# POST /cve/:id
 def test_create_cve_mismatch():
     """ the cve id in the json body does not match the cve id in the path parameter """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2000-0008_public.json') as json_file:
         data = json.load(json_file)
         res = requests.post(
             f'{env.AWG_BASE_URL}{CVE_URL}/{curr_cve_id}',
@@ -99,10 +96,9 @@ def test_create_cve_mismatch():
         response_contains_json(res, 'error', 'CVEID_MISMATCH')
 
 
-# POST /cve/:id
 def test_create_cve_reserved_state():
     """ the cve record cannot be created in the RESERVED state """
-    with open('./src/test/CVE-2017-5833_reserved.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2017-5833_reserved.json') as json_file:
         data = json.load(json_file)
         res = requests.post(
             f'{env.AWG_BASE_URL}{CVE_URL}/{reserved_cve_id}',
@@ -113,10 +109,9 @@ def test_create_cve_reserved_state():
         response_contains_json(res, 'error', 'CVE_UNSUPPORTED_RECORD_STATE')
 
 
-# POST /cve/:id
 def test_create_nonexistent_cveid():
     """ the cve record cannot be created because the CVE ID doesn't exist """
-    with open('./src/test/CVE-3000-0001_nonexistent.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-3000-0001_nonexistent.json') as json_file:
         data = json.load(json_file)
         res = requests.post(
             f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id_dne}',
@@ -127,10 +122,9 @@ def test_create_nonexistent_cveid():
         response_contains_json(res, 'error', 'CVEID_DNE')
 
 
-# POST /cve/:id
 def test_create_existent_cve():
     """ the cve record cannot be created because it already exists """
-    with open('./src/test/CVE-2021-0004_public.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2021-0004_public.json') as json_file:
         data = json.load(json_file)
         res = requests.post(
             f'{env.AWG_BASE_URL}{CVE_URL}/CVE-2021-0004',
@@ -141,9 +135,9 @@ def test_create_existent_cve():
         response_contains_json(res, 'error', 'CVE_RECORD_EXISTS')
 
 
-# PUT /cve/:id
+#### PUT /cve/:id ####
 def test_update_cve_record():
-    with open('./src/test/CVE-2021-0005_public.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2021-0005_public.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{curr_cve_id}',
@@ -153,10 +147,9 @@ def test_update_cve_record():
         ok_response_contains(res, curr_cve_id)
 
 
-# PUT /cve/:id
 def test_update_cve_mismatch():
     """ the cve id in the json body does not match the cve id in the path parameter """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2000-0008_public.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{curr_cve_id}',
@@ -167,10 +160,9 @@ def test_update_cve_mismatch():
         response_contains_json(res, 'error', 'CVEID_MISMATCH')
 
 
-# PUT /cve/:id
 def test_update_cve_reserved_state():
     """ the cve record cannot be changed to the RESERVED state """
-    with open('./src/test/CVE-2017-5833_reserved.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2017-5833_reserved.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{reserved_cve_id}',
@@ -181,10 +173,9 @@ def test_update_cve_reserved_state():
         response_contains_json(res, 'error', 'CVE_UNSUPPORTED_RECORD_STATE')
 
 
-# PUT /cve/:id
 def test_update_nonexistent_cve_id():
     """ the cve record cannot be updated because the CVE ID doesn't exist """
-    with open('./src/test/CVE-3000-0001_nonexistent.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-3000-0001_nonexistent.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id_dne}',
@@ -195,10 +186,9 @@ def test_update_nonexistent_cve_id():
         response_contains_json(res, 'error', 'CVEID_DNE')
 
 
-# PUT /cve/:id
 def test_update_nonexistent_cve():
     """ the cve record cannot be updated because it doesn't exist """
-    with open('./src/test/CVE-2009-0009_reject.json') as json_file:
+    with open('./src/test/cve_tests/cve_record_fixtures/CVE-2009-0009_reject.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{reject_cve_id}',
@@ -215,7 +205,7 @@ def test_update_nonexistent_cve():
 
 
 def update_cve(filename, cveid):
-    with open(f'./src/test/{filename}.json') as json_file:
+    with open(f'./src/test/cve_tests/cve_record_fixtures/{filename}.json') as json_file:
         data = json.load(json_file)
         res = requests.put(
             f'{env.AWG_BASE_URL}{CVE_URL}/{cveid}',
