+++ test-http/src/test/org_user_tests/org.py
@@ -15,7 +15,7 @@ from src.utils import (assert_contains, ok_response_contains,
 ORG_URL = '/api/org'
 
 
-# GET /org
+#### GET /org ####
 def test_get_all_orgs():
     """ secretariat users can request a list of all organizations """
     res = requests.get(
@@ -26,7 +26,7 @@ def test_get_all_orgs():
     assert len(json.loads(res.content.decode())['organizations']) >= 1
 
 
-# GET /org/:shortname
+#### GET /org/:shortname ####
 def test_get_mitre_cna():
     """ the cve services api contains the mitre cna """
     res = requests.get(
@@ -38,7 +38,7 @@ def test_get_mitre_cna():
     ok_response_contains_json(res, 'short_name', 'mitre')
 
 
-# GET /org/:shortname/id_quota
+#### GET /org/:shortname/id_quota ####
 def test_get_mitre_id_quota():
     """ the cve services api's mitre cna has a valid id quota """
     res = get_org_id_data('mitre')
@@ -57,6 +57,7 @@ def test_get_mitre_id_quota():
     assert quota == available + reserved
 
 
+#### PUT /org/:shortname ####
 def test_update_mitre_id_quota():
     """ a secretariat user can update its own ID quota """
     # NOTE: this test makes sure MITRE can reserve IDs for reservation tests
@@ -69,6 +70,7 @@ def test_update_mitre_id_quota():
         'mitre organization was successfully updated.')
 
 
+#### GET /org/:shortname/user/:username ####
 def test_get_mitre_demon_user():
     """ the user we're ... using, exists under the mitre cna """
     # although this is a bit tautological, it still exercises the endpoint
@@ -80,6 +82,7 @@ def test_get_mitre_demon_user():
     ok_response_contains_json(res, 'active', True)
 
 
+#### POST /org ####
 def test_post_new_org_empty_body():
     """ an empty new org body is invalid for name and short name """
     res = requests.post(
@@ -155,6 +158,7 @@ def test_post_new_org_duplicate_parameter():
     response_contains_json(res, 'message', 'Parameters were invalid')
 
 
+#### POST /org/:shortname/user ####
 def test_post_new_org_user():
     """ secretariat user can create a new user without active roles """
     uid = str(uuid.uuid4())
@@ -217,6 +221,7 @@ def test_post_new_org_user_duplicate():
         f"The user \'{env.AWG_USER_NAME}\' already exists.")
 
 
+#### PUT /org/:shortname ####
 def test_put_update_org_that_does_not_exist():
     """ cve services api will not update orgs that don't exist """
     uid = str(uuid.uuid4())
@@ -245,6 +250,7 @@ def test_put_update_mitre_org_nonexistent_user():
     response_contains_json(res, 'message', 'Unauthorized')
 
 
+#### PUT /org/:shortname/user/:username ####
 def test_put_update_user_username():
     """ services api allows user usernames to be updated by secretariat """
     org, user = create_new_user_with_new_org_by_uuid()
@@ -392,6 +398,7 @@ def test_put_update_user_remove_admin_role():
     response_contains(res, 'User role does not exist.')
 
 
+#### PUT /org/:shortname/user/:username/reset_secret ####
 def test_put_update_user_reset_secret():
     """ services api allows the secretariat to reset user secrets """
     org, user = create_new_user_with_new_org_by_uuid()
