+++ test-http/src/test/org_user_tests/org_as_org_admin.py
@@ -6,13 +6,13 @@ import json
 import requests
 import uuid
 from src import env, utils
-from src.test.org import (ORG_URL, create_new_user_with_new_org_by_uuid,
+from src.test.org_user_tests.org import (ORG_URL, create_new_user_with_new_org_by_uuid,
                             create_new_user_with_new_org_by_shortname,
                             post_new_org_user, post_new_org)
 from src.utils import response_contains, response_contains_json
 
 
-# GET /org
+#### GET /org ####
 def test_org_admin_get_all_orgs(org_admin_headers):
     """ services api rejects requests for all orgs by non-secretariat users """
     res = requests.get(
@@ -23,7 +23,7 @@ def test_org_admin_get_all_orgs(org_admin_headers):
     response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
-# GET /org/:shortname
+#### GET /org/:shortname ####
 def test_org_admin_get_mitre_org(org_admin_headers):
     """ services api rejects requests for secretariat by non-secretariat users """
     res = requests.get(
@@ -34,7 +34,6 @@ def test_org_admin_get_mitre_org(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname
 def test_org_admin_get_another_org(org_admin_headers):
     """ services api rejects requests for any org by another org user """
     different_org = str(uuid.uuid4()) # name of an org
@@ -48,7 +47,6 @@ def test_org_admin_get_another_org(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname
 def test_org_admin_get_own_org(org_admin_headers):
     """ services api allows org admins to get their own org's document """
     org = org_admin_headers["CVE-API-ORG"]
@@ -60,7 +58,7 @@ def test_org_admin_get_own_org(org_admin_headers):
     response_contains(res, org_admin_headers['CVE-API-ORG'])
 
 
-# GET /org/:shortname/id_quota
+#### GET /org/:shortname/id_quota ####
 def test_org_admin_get_secretariat_id_quota_info(org_admin_headers):
     """ services api rejects requests for secretariat by non-secretariat users """
     res = requests.get(
@@ -71,7 +69,6 @@ def test_org_admin_get_secretariat_id_quota_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/id_quota
 def test_org_admin_get_another_org_id_quota_info(org_admin_headers):
     """ services api rejects requests for any org by another org user """
     different_org = str(uuid.uuid4()) # name of an org
@@ -85,7 +82,6 @@ def test_org_admin_get_another_org_id_quota_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/id_quota
 def test_org_admin_get_own_id_quota_info(org_admin_headers):
     """ services api allows org admins to get info about their org's quota """
     org = org_admin_headers["CVE-API-ORG"]
@@ -99,7 +95,7 @@ def test_org_admin_get_own_id_quota_info(org_admin_headers):
     assert id_quota <= 100000
 
 
-# GET /org/:shortname/user/:username
+#### GET /org/:shortname/user/:username ####
 def test_org_admin_get_mitre_user_info(org_admin_headers):
     """ services api prevents org users from viewing secretariat user info """
     res = requests.get(
@@ -110,7 +106,6 @@ def test_org_admin_get_mitre_user_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/user/:username
 def test_org_admin_get_another_org_user_info(org_admin_headers):
     """ services api prevents org admin users from viewing another org user's info """
     org, user = create_new_user_with_new_org_by_uuid()
@@ -122,7 +117,6 @@ def test_org_admin_get_another_org_user_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/user/:username
 def test_org_admin_get_own_user_info(org_admin_headers):
     """ services api allows org admin to get its own user info """
     org = org_admin_headers['CVE-API-ORG']
@@ -135,7 +129,7 @@ def test_org_admin_get_own_user_info(org_admin_headers):
     response_contains(res, user)
 
 
-# GET /org/:shortname/users
+#### GET /org/:shortname/users ####
 def test_org_admin_get_mitre_users_info(org_admin_headers):
     """ services api prevents org users from viewing secretariat users info """
     res = requests.get(
@@ -146,7 +140,6 @@ def test_org_admin_get_mitre_users_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/users
 def test_org_admin_get_another_org_users_info(org_admin_headers):
     """ services api prevents org admin users from viewing all other org's user info """
     org = str(uuid.uuid4())
@@ -160,7 +153,6 @@ def test_org_admin_get_another_org_users_info(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_ORG_OR_SECRETARIAT')
 
 
-# GET /org/:shortname/users
 def test_org_admin_get_own_users_info(org_admin_headers):
     """ services api allows org admin to get its own users info """
     org = org_admin_headers['CVE-API-ORG']
@@ -174,7 +166,7 @@ def test_org_admin_get_own_users_info(org_admin_headers):
     response_contains(res, user)
 
 
-# POST /org
+#### POST /org ####
 def test_org_admin_cannot_create_another_org(org_admin_headers):
     """ services api does not allow org admins to create other orgs """
     res = requests.post(
@@ -186,7 +178,6 @@ def test_org_admin_cannot_create_another_org(org_admin_headers):
     response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
-# PUT /org/:shortname
 def test_org_admin_cannot_update_org(org_admin_headers):
     """ services api does not allow org admins to update their own orgs """
     res = requests.post(
@@ -198,7 +189,7 @@ def test_org_admin_cannot_update_org(org_admin_headers):
     response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
-# POST /org/:shortname/user
+#### POST /org/:shortname/user ####
 def test_org_admin_cannot_create_user_for_another_org(org_admin_headers):
     """ services api prevents org admins from creating a user with conflicts in the organization the user belongs to (org in path is diff from org in json body)  """
     org = str(uuid.uuid4())
@@ -213,7 +204,6 @@ def test_org_admin_cannot_create_user_for_another_org(org_admin_headers):
     response_contains_json(res, 'error', 'SHORTNAME_MISMATCH')
 
 
-# POST /org/:shortname/user
 def test_org_admin_cannot_create_user_for_another_org(org_admin_headers):
     """ services api prevents org admins from creating users for other orgs """
     org = str(uuid.uuid4())
@@ -228,7 +218,6 @@ def test_org_admin_cannot_create_user_for_another_org(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_ORG_ADMIN_OR_SECRETARIAT')
 
 
-# POST /org/:shortname/user
 def test_org_admin_cannot_create_existen_user(org_admin_headers):
     """ services api prevents org admins from creating existing users """
     user = str(uuid.uuid4())
@@ -244,7 +233,7 @@ def test_org_admin_cannot_create_existen_user(org_admin_headers):
     response_contains_json(res, 'error', 'USER_EXISTS')
 
 
-# PUT /org/:shortname/user/:username
+#### PUT /org/:shortname/user/:username ####
 def test_org_admin_cannot_update_user_org_dne(org_admin_headers):
     """ services api prevents org admins from updating a user from an org that doesn't exist """
     user = org_admin_headers['CVE-API-USER']
@@ -257,7 +246,6 @@ def test_org_admin_cannot_update_user_org_dne(org_admin_headers):
     response_contains_json(res, 'error', 'ORG_DNE_PARAM')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_user_dne(org_admin_headers):
     """ services api prevents org admins from updating a user that doesn't exist """
     user = str(uuid.uuid4())
@@ -270,7 +258,6 @@ def test_org_admin_cannot_update_user_dne(org_admin_headers):
     response_contains_json(res, 'error', 'USER_DNE')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_user_for_another_org(org_admin_headers):
     """ services api prevents org admins from updating a user from a diff org """
     org, user = create_new_user_with_new_org_by_uuid()
@@ -282,7 +269,6 @@ def test_org_admin_cannot_update_user_for_another_org(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_USER_OR_SECRETARIAT')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_user_new_shortname_dne(org_admin_headers):
     """ services api prevents org admins from updating a user's org that doesn't exist """
     org = org_admin_headers['CVE-API-ORG']
@@ -296,7 +282,6 @@ def test_org_admin_cannot_update_user_new_shortname_dne(org_admin_headers):
     response_contains_json(res, 'error', 'ORG_DNE')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_duplicate_user_with_new_shortname_and_username(org_admin_headers):
     """ services api prevents org admins from updating a user's org and username if that user already exist """
     org1 = org_admin_headers['CVE-API-ORG']
@@ -310,7 +295,6 @@ def test_org_admin_cannot_update_duplicate_user_with_new_shortname_and_username(
     response_contains_json(res, 'error', 'DUPLICATE_USERNAME')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_duplicate_user_with_new_username(org_admin_headers):
     """ services api prevents org admins from updating a user's username if that user already exist """
     org = org_admin_headers['CVE-API-ORG']
@@ -326,7 +310,6 @@ def test_org_admin_cannot_update_duplicate_user_with_new_username(org_admin_head
     response_contains_json(res, 'error', 'DUPLICATE_USERNAME')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_cannot_update_duplicate_user_with_new_shortname(org_admin_headers):
     """ services api prevents org admins from updating a user's org if that user already exist """
     user = org_admin_headers['CVE-API-USER']
@@ -341,7 +324,6 @@ def test_org_admin_cannot_update_duplicate_user_with_new_shortname(org_admin_hea
     response_contains_json(res, 'error', 'DUPLICATE_USERNAME')
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_update_same_org_user_state_sn_un(org_admin_headers):
     """  allows admin users to update a user's active state, org shortname, and user username """
     org = org_admin_headers['CVE-API-ORG']
@@ -362,7 +344,6 @@ def test_org_admin_update_same_org_user_state_sn_un(org_admin_headers):
     assert json.loads(res.content.decode())['updated']['username'] is not None
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_update_same_org_user_roles_name(org_admin_headers):
     """  allows admin users to update a user's name, add role, and remove role """
     org = org_admin_headers['CVE-API-ORG']
@@ -393,7 +374,6 @@ def test_org_admin_update_same_org_user_roles_name(org_admin_headers):
     assert json.loads(res.content.decode())['updated']['name']['surname'] == 's'
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_update_own_user_state_sn_un(org_admin_headers):
     """  allows admin users to update its own active state, org shortname, and user username """
     org = org_admin_headers['CVE-API-ORG']
@@ -412,7 +392,6 @@ def test_org_admin_update_own_user_state_sn_un(org_admin_headers):
     assert json.loads(res.content.decode())['updated']['username'] is not None
 
 
-# PUT /org/:shortname/user/:username
 def test_org_admin_update_own_user_roles_name(org_admin_headers):
     """  allows admin users to update its own name and remove role """ 
     org = org_admin_headers['CVE-API-ORG']
@@ -447,7 +426,7 @@ def test_org_admin_update_own_user_roles_name(org_admin_headers):
     assert json.loads(res.content.decode())['updated']['name']['surname'] == 's'
 
 
-# PUT /org/:shortname/user/:username/reset_secret
+#### PUT /org/:shortname/user/:username/reset_secret ####
 def test_org_admin_reset_secret_org_dne(org_admin_headers):
     org = str(uuid.uuid4())
     user = str(uuid.uuid4())
@@ -459,7 +438,6 @@ def test_org_admin_reset_secret_org_dne(org_admin_headers):
     response_contains_json(res, 'error', 'ORG_DNE_PARAM')
 
 
-# PUT /org/:shortname/user/:username/reset_secret
 def test_org_admin_reset_secret_org_dne(org_admin_headers):
     org = org_admin_headers['CVE-API-ORG']
     user = str(uuid.uuid4())
@@ -471,7 +449,6 @@ def test_org_admin_reset_secret_org_dne(org_admin_headers):
     response_contains_json(res, 'error', 'USER_DNE')
 
 
-# PUT /org/:shortname/user/:username/reset_secret
 def test_org_admin_reset_diff_org_secret(org_admin_headers):
     """ services api prevents admin users to reset the secret of users of different org"""
     org, user = create_new_user_with_new_org_by_uuid()
@@ -483,7 +460,6 @@ def test_org_admin_reset_diff_org_secret(org_admin_headers):
     response_contains_json(res, 'error', 'NOT_SAME_USER_OR_SECRETARIAT')
 
 
-# PUT /org/:shortname/user/:username/reset_secret
 def test_org_admin_reset_same_org_secret(org_admin_headers):
     """ services api allows admin users to reset the secret of users of same org"""
     org = org_admin_headers['CVE-API-ORG']
@@ -498,7 +474,6 @@ def test_org_admin_reset_same_org_secret(org_admin_headers):
     response_contains(res, 'API-secret')
 
 
-# PUT /org/:shortname/user/:username/reset_secret
 def test_org_admin_reset_own_secret(org_admin_headers):
     """ services api allows admin users to reset their own secret """
     org = org_admin_headers['CVE-API-ORG']
