+++ test-http/src/test/org_user_tests/regular_users.py
@@ -6,7 +6,7 @@ import json
 import requests
 import uuid
 from src import env, utils
-from src.test.org import (ORG_URL, create_new_user_with_new_org_by_uuid,
+from src.test.org_user_tests.org import (ORG_URL, create_new_user_with_new_org_by_uuid,
                             create_new_user_with_new_org_by_shortname,
                             post_new_org_user, post_new_org)
 from src.utils import (assert_contains, ok_response_contains,
@@ -350,7 +350,7 @@ def test_regular_user_cannot_reset_secret_user_dne(reg_user_headers):
     response_contains_json(res, 'error', 'USER_DNE')
 
 
-#### Post /org/ ####
+#### POST /org ####
 def test_reg_user_cannot_post_new_org(reg_user_headers):
     """ regular users cannot create new org """
     res = requests.post(
